updated files about postToFirebase
